+++ src/main/java/au/gov/nehta/model/clinical/etp/common/item/Logo.java
@@ -1,6 +1,7 @@
 package au.gov.nehta.model.clinical.etp.common.item;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
@@ -10,14 +11,14 @@ public class Logo extends AttachedMedia {
     public final Set<String> ALLOWED_MIME_TYPES =
             new HashSet<>(Arrays.asList("image/jpeg", "image/png", "image/gif"));
 
-    public Logo(File file) {
+    public Logo(File file) throws IOException {
         super(file);
 
         if (!ALLOWED_MIME_TYPES.contains(getMediaType()))
             throw new IllegalArgumentException("Invalid logo media type. Only png, gif, jpg allowed");
     }
 
-    public Logo(File logo, IntegerityCheckType checkType) {
+    public Logo(File logo, IntegerityCheckType checkType) throws IOException {
         super(logo, checkType);
         if (!ALLOWED_MIME_TYPES.contains(getMediaType()))
             throw new IllegalArgumentException("Invalid logo media type. Only png, gif, jpg allowed");