+++ src/test/java/au/gov/nehta/cda/sl/SpecialistLetterMaxTest.java
@@ -48,6 +48,7 @@ import javax.xml.bind.JAXBException;
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
@@ -76,13 +77,13 @@ public class SpecialistLetterMaxTest extends Base {
             File f = new File(DOCUMENT_FILE_NAME);
             Assert.assertTrue(f.exists());
             Assert.assertTrue(f.length() > 0L);
-        } catch (SchematronValidationException | ParserConfigurationException | JAXBException | FileNotFoundException e) {
+        } catch (SchematronValidationException | ParserConfigurationException | JAXBException | IOException e) {
             e.printStackTrace();
         }
     }
 
     private void generateMax()
-            throws SchematronValidationException, JAXBException, ParserConfigurationException, FileNotFoundException {
+        throws SchematronValidationException, JAXBException, ParserConfigurationException, IOException {
 
         // Prepare Specialist Letter Context
         ParticipationServiceProvider referrer = getCurrentServiceProviderIndividual();
@@ -229,7 +230,7 @@ public class SpecialistLetterMaxTest extends Base {
     private DiagnosticInvestigations getDiagnosticInvestigations(boolean showPathologyTestResults,
                                                                  boolean showImagingExaminationResults, boolean showRequestedServices,
                                                                  boolean showOtherTestResults)
-            throws FileNotFoundException {
+        throws IOException {
         DiagnosticInvestigations diagnosticInvestigations = new DiagnosticInvestigationsImpl();
         //Pathology Test Results
         if (showPathologyTestResults) {
@@ -262,7 +263,7 @@ public class SpecialistLetterMaxTest extends Base {
         return diagnosticInvestigations;
     }
 
-    private OtherTestResult getOtherTestResultWithReportFile() {
+    private OtherTestResult getOtherTestResultWithReportFile() throws IOException {
         AttachedMedia reportFile = TestHelper.getAttachedMediaPDF("radiologyreport.pdf");
         return new OtherTestResultImpl(new CodeImpl() {{
             setOriginalText("Report Name (with attachment)");
@@ -520,7 +521,7 @@ public class SpecialistLetterMaxTest extends Base {
         return referenceRangeDetails;
     }
 
-    private static AnatomicalSite getAnatomicalSite(String s) throws FileNotFoundException {
+    private static AnatomicalSite getAnatomicalSite(String s) throws IOException {
         AnatomicalSite anatomicalSite = new AnatomicalSiteImpl();
         anatomicalSite.setAnatomicalLocationImages(new ArrayList<AttachedMedia>() {{
             add(getAttachedMedia(s, Optional.of(ATTACHMENTS_DIR)));
@@ -536,7 +537,7 @@ public class SpecialistLetterMaxTest extends Base {
         return anatomicalSite;
     }
 
-    private ExaminationRequestDetails getExaminationRequestDetails() throws FileNotFoundException {
+    private ExaminationRequestDetails getExaminationRequestDetails() throws IOException {
         // Image Details
         ImageDetails imageDetail = new ImageDetailsImpl();
         imageDetail.setSubjectPosition("Subject Pos");
@@ -570,7 +571,7 @@ public class SpecialistLetterMaxTest extends Base {
         return examinationRequestDetails;
     }
 
-    private ImagingExaminationResult getImagingExaminationResult() throws FileNotFoundException {
+    private ImagingExaminationResult getImagingExaminationResult() throws IOException {
 
         List<AnatomicalSite> anatomicalSites = new ArrayList<>();
         anatomicalSites.add(getAnatomicalSite("ImagingExaminationResult"));